dto/src/test/java/com/github/xavierlepretre/tmdb/model/movie/GenreDTOTest.java
package com.github.xavierlepretre.tmdb.model.movie;

import com.fasterxml.jackson.databind.JsonMappingException;
import com.fasterxml.jackson.databind.ObjectMapper;

import org.junit.Before;
import org.junit.Rule;
import org.junit.Test;
import org.junit.rules.ExpectedException;

import static org.fest.assertions.api.Assertions.assertThat;

public class GenreDTOTest
{
    private ObjectMapper mapper;

    @Rule
    public ExpectedException thrown= ExpectedException.none();

    @Before public void setUp()
    {
        this.mapper = new ObjectMapper();
    }

    @Test
    public void canDeserialise() throws Exception
    {
        GenreDTO dto = mapper.readValue(getClass().getResourceAsStream("genre_dto_1.json"), GenreDTO.class);
        assertThat(dto.getId()).isEqualTo(new GenreId(12));
        assertThat(dto.getName()).isEqualTo("Adventure");
    }

    @Test
    public void cannotDeserialiseMissingId() throws Exception
    {
        thrown.expect(JsonMappingException.class);
        thrown.expectMessage("id");
        mapper.readValue(getClass().getResourceAsStream("genre_dto_1_missing_id.json"), GenreDTO.class);
    }

    @Test
    public void cannotDeserialiseMissingName() throws Exception
    {
        thrown.expect(JsonMappingException.class);
        thrown.expectMessage("name");
        mapper.readValue(getClass().getResourceAsStream("genre_dto_1_missing_name.json"), GenreDTO.class);
    }
}
